+++ prefs.py
@@ -7,6 +7,7 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
         default = 1.0, min = 0.01, max = 100.0, soft_min = 0.1, soft_max = 10.0, step = 1, precision = 2)
     sensitivityStrafe: bpy.props.FloatProperty(name = "Strafe Sensitivity", description = "Default mouse speed multiplier when mouse strafing. Use Alt + Mousewheel to customize this value on a per-blend-file basis", \
         default = 1.0, min = 0.01, max = 100.0, soft_min = 0.1, soft_max = 10.0, step = 1, precision = 2)
+    strafeGears: bpy.props.FloatVectorProperty(name = "Gears", description = "Set additional strafe multipliers to cycle through with G/Shift-G. Entries set to 0 are ignored", size = 7, default = (0.1, 1.0, 5.0, 0, 0, 0, 0), min = 0.0, max = 1000.0, soft_min = 0.0, soft_max = 100.0, step = 5, precision = 2)
 
     invertMouse: bpy.props.BoolProperty(name = "Invert Mouse Vertically", description = "Invert effect of vertical mouse movement when looking around", default = True)
     invertMouseX: bpy.props.BoolProperty(name = "Invert Mouse Horizontally", description = "Invert effect of horizontal mouse movement when looking around", default = False)
@@ -53,8 +54,9 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
     keyRight: bpy.props.StringProperty(name = "Move Right", description = "Press this key to stafe the camera to the right (must be upper-case)", default = "D")
     keyUp: bpy.props.StringProperty(name = "Move Up", description = "Press this key to strafe the camera upwards (must be upper-case)", default = "E")
     keyDown: bpy.props.StringProperty(name = "Move Down", description = "Press this key to strafe the camera downwards (must be upper-case)", default = "Q")
-
     keyResetRoll: bpy.props.StringProperty(name = "Reset Roll", description = "Press this key to reset camera roll (must be upper-case)", default = "R")
+
+    keyCycleGears: bpy.props.StringProperty(name = "Cycle Gear", description = "Press this key to cycle through strafing gears (must be upper-case)", default = "G")
     keyRelocatePivot: bpy.props.StringProperty(name = "Relocate Pivot", description = "Press this key to relocate camera pivot to the nearest surface in the center of the view. You can also hold it for one second to toggle Automatically Relocate Pivot on and off (must be upper-case)", default = "C")
     keyLoadCameraState: bpy.props.StringProperty(name = "Load Camera State", description = "Press this key and then one of the number keys [0-9] to load the camera state in that slot without the need to quickly press the number key twice (must be upper-case)", default = "T")
 
@@ -68,6 +70,7 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
         row = box.row()
         row.prop(self, "sensitivityPan")
         row.prop(self, "sensitivityStrafe")
+        box.row().prop(self, "strafeGears")
         
         row = box.row()
         row.prop(self, "invertMouse")
@@ -97,11 +100,11 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
 
         row = box.row()
         row.prop(self, "showCrosshair")
-        row.prop(self, "debug")
+        row.prop(self, "leaveFOV")
 
         row = box.row()
         row.prop(self, "toggleMode")
-        row.prop(self, "leaveFOV")
+        row.prop(self, "debug")
 
         box = layout.box()
 
@@ -120,5 +123,6 @@ class MouseStrafingPreferences(bpy.types.AddonPreferences):
         box.row().prop(self, "keyUp")
         box.row().prop(self, "keyDown")
         box.row().prop(self, "keyResetRoll")
+        box.row().prop(self, "keyCycleGears")
         box.row().prop(self, "keyRelocatePivot")
         box.row().prop(self, "keyLoadCameraState")